Denoise a SALT Observatory spectrum via quadratic trend filtering and compute uncertainties via a bootstrap

Denoise a SALT Observatory spectrum via quadratic trend filtering and compute uncertainties via a bootstrap

denoise_polarized_spectrum(
  wavelength,
  stokes,
  variances,
  masks,
  break_at = 10,
  min_pix_segment = 10,
  compute_uncertainties = FALSE,
  mc_cores = parallel::detectCores(),
  ...
)

Arguments

stokes
Spectropolarimetric measurements in a Stokes parametrization, passed as a 3-column matrix, with the columns corresponding to the I, Q, U Stokes parameters, respectively.

variances
Measurement variances, in a matrix with dimensions matching those of stokes.

masks
Pixel masks, in a matrix with dimensions matching those of stokes. Nonzero elements flag bad pixels.

break_at
A free parameter that controls the segmentation of a spectrum. More precisely, break_at is the minimum number of consecutively-masked spectral pixels that will trigger a break in the spectrum. Defaults to break_at = 10.

min_pix_segment
After the segmentation procedure is complete, it is advisable to examine the resulting segments to ensure that each is sufficiently long for its own denoising analysis. In particular, we discard any segments that have less than min_pix_segment unmasked spectral pixels. Defaults to min_pix_segment = 10.

compute_uncertainties
(Boolean) If TRUE then variability bands are computed for each of the denoised normalized Stokes spectra, via a parametric bootstrap algorithm.

...
Additional named arguments to be passed to trendfiltering::sure_trendfilter() or trendfiltering::bootstrap_trendfilter().

Value

An object of class 'polarized_spectrum_denoised'. This is a list with the following elements:

References

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

Examples

data(SALT_spectrum)

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
)

stokes <- as_tibble(sci$imDat) %>%
  rename_with(function(.cols) c("I", "Q", "U"))
variances <- as_tibble(var$imDat) %>%
  select(1:3) %>%
  rename_with(function(.cols) c("I_vars", "Q_vars", "U_vars"))
masks <- as_tibble(bpm$imDat) %>%
  rename_with(function(.cols) c("I_mask", "Q_mask", "U_mask"))