Denoise a polarized spectrum via quadratic trend filtering

The denoise_spectrum() function uses quadratic trend filtering, tuned by Stein's unbiased risk estimate, to optimally denoise a polarized spectrum in each of the \(I\), \(Q\), \(U\) Stokes parameters (sometimes denoted \(S_0\), \(S_1\), \(S_2\)), which then also leads to estimates for the normalized Stokes parameters \(Q/I\) and \(U/I\). Setting compute_uncertainties = TRUE generates a bootstrap ensemble of denoised spectra for each Stokes parameter, which allows variability bands to be computed for each denoised spectrum by then calling variability_bands() on the 'polarized_spectrum' object produced by the denoise_spectrum() function call.

denoise_spectrum(
  wavelength,
  flux,
  variances,
  masks,
  break_at = 10L,
  min_pix_segment = 10L,
  compute_uncertainties = FALSE,
  B = 100L,
  mc_cores = parallel::detectCores() - 4,
  validation_args
)

Arguments

wavelength
Vector of wavelength measurements.

flux
Spectropolarimetric measurements, passed as a 3-column tibble, data frame, or matrix, with the columns corresponding to the \(I\), \(Q\), \(U\) Stokes parameters, respectively.

variances
Measurement variances, in a tibble, data frame, or matrix with dimensions matching those of flux.

masks
Pixel masks, in a tibble, data frame, or matrix with dimensions matching those of flux. Nonzero elements flag bad pixels.

break_at
The minimum number of consecutively-masked spectral pixels that will trigger a break in the spectrum. Defaults to break_at = 10.

min_pix_segment
After the segmentation procedure is complete, the resulting segments are examined to ensure that each is sufficiently long for a non-trivial/well-defined denoising analysis. In particular, any segment that has less than min_pix_segment unmasked spectral pixels is discarded. Defaults to min_pix_segment = 10.

compute_uncertainties
(Boolean) If TRUE, then bootstrap ensembles are created for each denoised spectrum via bootstrap_trendfilter(). The ensembles are stored in the function output so that variability bands of any level can quickly be computed by calls to variability_bands() without redundant overhead (see examples). Defaults to compute_uncertainties = FALSE.

B
If compute_uncertainties = TRUE, B is the number of bootstrap samples in the ensembles used to estimate the uncertainty in the denoiser. Defaults to B = 100.

mc_cores
Multi-core computing using the parallel package: The number of cores to use. Defaults to the number of cores detected on the machine, minus 4.

validation_args
(Optional) A named list of arguments to be passed to sure_trendfilter(). The evaluation grid argument x_eval of sure_trendfilter() defaults to the observed wavelength grid. The arguments x, y, weights, and k cannot be overridden.

Value

An object of class 'polarized_spectrum'. This is a list with the following elements:

n_segments

The number of segments the spectrum was broken into by break_spectrum().

data

The original data set, as a list of n_segments tibbles. Each tibble contains the observed wavelengths (with the superset of masked wavelengths removed), the Stokes \(I\), \(Q\), \(U\) flux measurements, and the Stokes \(I\), \(Q\), \(U\) flux measurement variances.

denoised_signals

The set of denoised spectra, as a list of n_segments tibbles. Each tibble contains the wavelength evaluation grid for its respective spectrum segment and all of the denoised signal estimates: \(I\), \(Q\), \(U\), \(Q/I\), \(U/I\).

ensembles

If compute_uncertainties = TRUE, a list of bootstrap ensembles for the denoised \(I\), \(Q\), and \(U\) ensembles, respectively. Each ensemble is returned as an \(n \times B\) matrix, less any columns from post-hoc pruning (see the bootstrap_trendfilter() documentation). If compute_uncertainties = FALSE, this will return NULL.

I_analysis_summary

Technical summary of the denoising analysis of the \(I\) Stokes parameter.

Q_analysis_summary

Same as above, but for the \(Q\) Stokes parameter.

U_analysis_summary

Same as above, but for the \(U\) Stokes parameter.

References

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

See also

Examples

# Any SALT Observatory spectrum can be read into R from its FITS file
# using the "FITSio" R package, as below. For convenience, here we've stored
# the `sci`, `var`, and `bpm` data structures for a Wolf-Rayet stellar
# spectrum in an R data file within this package, so we can source them using
# a simple call to `data()`.

if (FALSE) {
install.packages("FITSio")
path_to_FITS_file <- "<your_local_path_to_the_FITS_file>"
file.name <- "WR006_c1_12345678_stokes.fits"
sci <- FITSio::readFITS(paste0(path_to_FITS_file, file.name), hdu = 1)
var <- FITSio::readFITS(paste0(path_to_FITS_file, file.name), hdu = 2)
bpm <- FITSio::readFITS(paste0(path_to_FITS_file, file.name), hdu = 4)
}

####

data(polarized_spectrum_WR_star)
library(dplyr)

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
)

flux <- as_tibble(sci$imDat)
variances <- as_tibble(var$imDat) %>% select(1:3)
masks <- as_tibble(bpm$imDat)

spec_denoised <- denoise_spectrum(
  wavelength,
  flux,
  variances,
  masks,
  compute_uncertainties = TRUE
)