Quantify the statistical uncertainty in a denoised spectrum by computing bootstrap variability bands

Quantify the statistical uncertainty in a denoised spectrum by computing bootstrap variability bands

variability_bands(obj, param, level = 0.95)

Arguments

obj
An object of class "polarized_spectrum" produced by denoise_spectrum().

param
A string specifying which spectrum to compute variability bands for. One of c("I","Q","U","Q_norm","U_norm").

level
The level of the pointwise variability bands. Defaults to level = 0.95.

Value

A list of tibbles, with length equal to obj$n_segments and each tibble with column set c("wavelength","bootstrap_lower_band","bootstrap_upper_band").

Examples

data(polarized_spectrum_WR_star)

suppressMessages(library(dplyr))

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
)

flux <- as_tibble(sci$imDat)
variances <- as_tibble(var$imDat) %>% select(1:3)
masks <- as_tibble(bpm$imDat)

# Not running the rest to save time during development commits
if (FALSE) {
spec_denoised <- denoise_spectrum(
  wavelength,
  flux,
  variances,
  masks,
  compute_uncertainties = TRUE
)

bands <- variability_bands(spec_denoised, param = "Q_norm", level = 0.95)}