Break a spectrum into segments when one or more wide masks exist

Break a spectrum into segments when one or more wide masks (i.e. sequences of many consecutively-masked pixels) exist. The spectrum can be polarized or unpolarized (see examples). When the spectrum is polarized, the set of masked spectral pixels used to define the spectrum's segmentation is taken to be the set of pixels that are masked for any of the \(I\), \(Q\), \(U\) Stokes parameters. Note that denoise_spectrum() calls break_spectrum() internally. So while break_spectrum() is an exported function, it typically does not need to be called directly.

break_spectrum(df_full, break_at = 10, min_pix_segment = 10)

Arguments

df_full
A tibble that includes a column titled "wavelength" and one or more columns with names ending in "mask", e.g. c("I_mask", "Q_mask", "U_mask"). The values of the mask columns are equal to 0 (or FALSE) for unmasked pixels and nonzero (or TRUE) for masked pixels.

break_at
The minimum number of consecutively-masked spectral pixels that will trigger a break in the spectrum. Defaults to break_at = 10.

min_pix_segment
After the segmentation procedure is complete, the resulting segments are examined to ensure that each is sufficiently long for a non-trivial/well-defined denoising analysis. In particular, any segment that has less than min_pix_segment unmasked spectral pixels is discarded. Defaults to min_pix_segment = 10.

Value

A list of tibbles that define the segmented spectrum. Each tibble in the list has the same format as the input tibble df_full, minus the mask columns.

See also

Examples

data(polarized_spectrum_WR_star)

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
) %>%
  tibble::as_tibble_col(column_name = "wavelength")

stokes <- as_tibble(sci$imDat) %>%
  rename_with(function(.cols) c("I", "Q", "U"))

variances <- as_tibble(var$imDat) %>%
  select(1:3) %>%
  rename_with(function(.cols) c("I_vars", "Q_vars", "U_vars"))

masks <- as_tibble(bpm$imDat) %>%
  rename_with(function(.cols) c("I_mask", "Q_mask", "U_mask"))

df_full <- bind_cols(
  wavelength,
  stokes,
  variances,
  masks
)

df_list <- break_spectrum(df_full)