Denoise a polarized spectrum

Denoise a polarized spectrum.

denoise_polarized_spectrum(
  wavelength,
  stokes,
  variances,
  masks,
  break_at = 10,
  min_pix_segment = 10,
  compute_uncertainties = FALSE,
  sure_args,
  bootstrap_args
)

Arguments

wavelength
Vector of wavelength measurements.

stokes
Polarized spectrum measurements, passed as a 3-column tibble, data frame, or matrix, with the columns corresponding to the Stokes parameters I, Q, and U, respectively.

variances
Measurement variances, in a tibble, data frame, or matrix with dimensions matching those of stokes.

masks
Pixel masks, in a tibble, data frame, or matrix with dimensions matching those of stokes. Nonzero elements flag bad pixels.

break_at
A free parameter that controls the segmentation of a spectrum. More precisely, break_at is the minimum number of consecutively-masked spectral pixels that will trigger a break in the spectrum. Defaults to break_at = 10.

min_pix_segment
After the segmentation procedure is complete, the resulting segments are examined to ensure that each is sufficiently long for an independent denoising analysis. In particular, any segment that has less than min_pix_segment unmasked spectral pixels is discarded. Defaults to min_pix_segment = 10.

compute_uncertainties
(Boolean) If TRUE, then variability bands are computed for each of the denoised normalized Stokes spectra, via a parametric bootstrap algorithm.

sure_args
Additional named arguments to be passed to sure_trendfilter().

bootstrap_args
Additional named arguments to be passed to bootstrap_trendfilter(). The argument bootstrap_algorithm = "parametric" is fixed due to the design of SALT spectra, and we don't allow this argument to be manually overridden. See the bootstrap_trendfilter() documentation and Politsch et al. (2020a) for details on why this particular bootstrap algorithm is appropriate for SALT spectra.

Value

An object of class 'polarized_spectrum'. This is a list with the following elements:

n_segments

The number of segments the spectrum was broken into.

denoised_spectra

A tibble containing denoised spectra for each of the 3 Stokes parameters, as well as Q/I and U/I. The full column set is c("wavelength","stokes_I","stokes_Q","stokes_U","Q_norm","U_norm").

stokes_I_ensemble

If compute_uncertainties = TRUE, the full bootstrap ensemble for the Stokes I parameter, as an \(n \times B\) matrix, less any columns from post-hoc pruning. If compute_uncertainties = FALSE, this will return NULL.

stokes_Q_ensemble

Same as above, but for the Q Stokes parameter.

stokes_U_ensemble

Same as above, but for the U Stokes parameter.

Q_norm_ensemble

Same as above, but for Q/I.

U_norm_ensemble

Same as above, but for U/I.

References

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

Examples

data(polarized_spectrum_WR_star)

suppressMessages(library(dplyr))

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
)

stokes <- as_tibble(sci$imDat)
variances <- as_tibble(var$imDat) %>% select(1:3)
masks <- as_tibble(bpm$imDat)

spec_denoised <- denoise_polarized_spectrum(wavelength, stokes, variances, masks)

#> Warning: all scheduled cores encountered errors in user code

#> Error in tf_obj[[X]][["x_eval"]]: subscript out of bounds